Initial commit
Old project that was originally part of a website made for a private Discord server. The website contained various features, including this project, which now has been extracted and made standalone. There has been multiple changes made to this version, such as translations to English, fixes related to deprecated functions/libraries, and a reimplementation of the authentication and login system. Despite these changes, the core functionality remains largely unchanged from the original project, and I intend to preserve this unless necessary changes are required.